backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
openai==1.40.0
groq>=0.9.0
pypdf>=4.2.0
